requirements.txt
python-telegram-bot==20.*
openpyxl
spacy
textacy